telegram-bot/requirements.txt
python-telegram-bot[webhooks]>=21.0
psycopg2-binary>=2.9
httpx>=0.27
bcrypt>=4.0
pydantic-settings>=2.0
python-dotenv>=1.0
uvicorn>=0.34
fastapi>=0.115
